8974适配

+++ tools/soc_msm8x26.sh
@@ -20,8 +20,10 @@ function soc_build_all()
 function soc_prebuild()
 {
 	echo $FUNCNAME
+	echo $DBG_PLATFORM
 	cd $DBG_SYSTEM_DIR
 
+if [ $DBG_PLATFORM = msm8226 ];then
 	rm -rf $DBG_SYSTEM_DIR/kernel/drivers/flyaudio
 	mkdir -p $DBG_SYSTEM_DIR/kernel/drivers/flyaudio
 	cp -ru $DBG_DRIVERS_PATH/build_in/*	        $DBG_SYSTEM_DIR/kernel/drivers/flyaudio/
@@ -30,13 +32,12 @@ function soc_prebuild()
 	cp -u $DBG_CORE_PATH/inc/cmn_func.h   $DBG_SYSTEM_DIR/kernel/drivers/flyaudio/
 	cp -u $DBG_CORE_PATH/inc/lidbg_def.h   $DBG_SYSTEM_DIR/kernel/drivers/flyaudio/
 
-
 	cp -r $RELEASE_REPOSITORY/driver/out $DBG_SYSTEM_DIR/out/target/product/$DBG_PLATFORM/system/lib/modules/
 	cp $RELEASE_REPOSITORY/driver/out/vold $DBG_SYSTEM_DIR/out/target/product/$DBG_PLATFORM/system/bin/
 	cp $RELEASE_REPOSITORY/app/FastBoot.apk $DBG_SYSTEM_DIR/out/target/product/$DBG_PLATFORM/system/app/
 	cp $RELEASE_REPOSITORY/app/FlyBootService.apk $DBG_SYSTEM_DIR/out/target/product/$DBG_PLATFORM/system/app/
 	cp $RELEASE_REPOSITORY/driver/out/lidbg_load $DBG_SYSTEM_DIR/out/target/product/$DBG_PLATFORM/system/bin/
-
+fi
 
 	if [[ $TARGET_PRODUCT = "" ]];then
 		source build/envsetup.sh&&choosecombo release $DBG_PLATFORM $SYSTEM_BUILD_TYPE
